requirements.txt
streamlit>=1.20
SQLAlchemy>=1.4
psycopg2-binary>=2.9   # for PostgreSQL (or use mysqlclient for MySQL)
pandas>=1.5
requests>=2.28
python-dotenv>=1.0
alembic>=1.12          # optional for migrations
pymysql 
sqlalchemy
plotly
openpyxl
xlsxwriter